+++ package.json
@@ -9,6 +9,7 @@
     "@cypress/react": "^5.12.4",
     "bulma": "^0.9.3",
     "classnames": "^2.3.1",
+    "lodash.debounce": "^4.0.8",
     "postcss": "^8.4.12",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
@@ -22,6 +23,7 @@
     "@mate-academy/scripts": "^1.2.8",
     "@mate-academy/students-ts-config": "*",
     "@mate-academy/stylelint-config": "*",
+    "@types/lodash.debounce": "^4.0.7",
     "@types/node": "^17.0.23",
     "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.14",
