+++ src/App.tsx
@@ -1,34 +1,77 @@
-import React, { useMemo, useState } from 'react';
+import React, { useCallback, useMemo, useState } from 'react';
+import debounce from 'lodash.debounce';
 
 import './App.scss';
 import { Dropdown } from './components';
 import { peopleFromServer } from './data/people';
 import { Person } from './types/Person';
 
-export const App: React.FC = () => {
-  const [onSelected, setOnSelected] = useState<Person | null>(null);
+type Props = {
+  delay: number,
+};
 
-  const selectedPerson = useMemo(() => {
-    if (onSelected) {
-      const { name, born, died } = onSelected;
+export const App: React.FC<Props> = ({ delay }) => {
+  const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
+  const [value, setValue] = useState('');
+  const [appliedValue, setAppliedValue] = useState('');
 
-      return `${name} (${born} - ${died})`;
-    }
+  const preparedPeople = useMemo(() => {
+    return peopleFromServer.filter(person => (
+      person.name.toLowerCase().includes(appliedValue)
+    )) || null;
+  }, [appliedValue, peopleFromServer]);
 
-    return 'No selected person';
-  }, [onSelected]);
+  const applyValue = useCallback(
+    debounce(setAppliedValue, delay),
+    [delay],
+  );
+
+  const handleChange = (
+    event: React.ChangeEvent<HTMLInputElement>,
+  ) => {
+    const normalisedEvent = event.target.value.toLowerCase();
+
+    setValue(event.target.value);
+    applyValue(normalisedEvent);
+  };
+
+  const handleClick = useCallback((
+    person: Person,
+    event: React.MouseEvent<HTMLAnchorElement, MouseEvent>,
+  ) => {
+    event.preventDefault();
+
+    setValue(person.name);
+    setSelectedPerson(person);
+    setAppliedValue('');
+  }, []);
 
   return (
     <main className="section">
       <h1 className="title">
-        {selectedPerson}
+        {selectedPerson
+          ? `${selectedPerson.name} (${selectedPerson.born} - ${selectedPerson.died})`
+          : 'No selected person'}
       </h1>
 
-      <Dropdown
-        people={peopleFromServer}
-        setOnSelected={setOnSelected}
-        delay={1000}
-      />
+      <div className="dropdown is-active">
+        <div className="dropdown-trigger">
+          <input
+            type="text"
+            placeholder="Enter a part of the name"
+            className="input"
+            value={value}
+            onChange={handleChange}
+          />
+        </div>
+
+        {appliedValue && (
+          <Dropdown
+            people={preparedPeople}
+            handleClick={handleClick}
+          />
+        )}
+      </div>
     </main>
   );
 };
