+++ src/index.tsx
@@ -2,4 +2,4 @@ import ReactDOM from 'react-dom';
 import 'bulma/css/bulma.css';
 import { App } from './App';
 
-ReactDOM.render(<App />, document.getElementById('root'));
+ReactDOM.render(<App delay={1000} />, document.getElementById('root'));
