+++ tsconfig.json
@@ -2,5 +2,5 @@
   "extends": "@mate-academy/students-ts-config",
   "include": [
     "src"
-  ]
+, "src/components/DropdownList/DropdownList.tsx"  ]
 }
